requirements.txt
Babel==2.2.0
Django==1.9.1
django-phonenumber-field==1.0.0
freezegun==0.3.5
phonenumbers==7.2.4
python-dateutil==2.4.2
pytz==2015.7
six==1.10.0

# for heroku
dj-database-url==0.3.0
gunicorn==19.4.5
psycopg2==2.5.4
